KAT-9 refactor: 結果画像 renderer を Core orchestrator + WMGP module layout に分離
