+++ apps/ledger/app/leagues/wmgp/league_module.rb
@@ -30,7 +30,7 @@ module Wmgp
 
     class Renderers
       def match_result_card
-        ::MatchExports::ResultCardRenderer
+        ::Wmgp::Renderers::MatchResultCard
       end
 
       def standings_table
