+++ apps/ledger/test/services/match_exports/result_card_renderer_test.rb
@@ -1,78 +1,6 @@
 require "test_helper"
 
 class MatchExports::ResultCardRendererTest < ActiveSupport::TestCase
-  test "versus header uses VS label" do
-    match = create_match_for_renderer_test!(home_name: "Alpha Team", away_name: "Beta Team")
-
-    html = MatchExports::ResultCardRenderer.new(match).send(:versus_html)
-
-    assert_includes html, ">VS<"
-    assert_not_includes html, ">対<"
-  end
-
-  test "renderer falls back to board scores when stored winners are stale" do
-    match = create_match_for_renderer_test!(home_name: "StardustHope", away_name: "ST | KnikxLaw")
-    round_one = match.rounds.create!(
-      number: 1,
-      home_team: match.home_team,
-      away_team: match.away_team,
-      result_status: "partial",
-      winner_team: nil
-    )
-    round_two = match.rounds.create!(
-      number: 2,
-      home_team: match.home_team,
-      away_team: match.away_team,
-      result_status: "confirmed",
-      winner_team: match.home_team
-    )
-
-    create_confirmed_board_result!(round: round_one, board_number: 1, score: [2, 0], winner_side: nil)
-    create_confirmed_board_result!(round: round_one, board_number: 2, score: [2, 0], winner_side: nil)
-    create_confirmed_board_result!(round: round_one, board_number: 3, score: [2, 0], winner_side: nil)
-    create_confirmed_board_result!(round: round_two, board_number: 1, score: [2, 1], winner_side: "home")
-    create_confirmed_board_result!(round: round_two, board_number: 2, score: [2, 1], winner_side: "home")
-    create_confirmed_board_result!(round: round_two, board_number: 3, score: [1, 2], winner_side: "away")
-
-    match.create_match_result!(
-      home_round_wins: 1,
-      away_round_wins: 0,
-      winner_team: match.home_team,
-      result_status: "partial",
-      decision_type: "normal"
-    )
-
-    renderer = MatchExports::ResultCardRenderer.new(match.reload)
-
-    round_one_html = renderer.send(:round_result_html, match.rounds.find_by!(number: 1))
-    footer_html = renderer.send(:footer_html)
-
-    assert_includes round_one_html, '<span class="round-result__label--win">W</span>'
-    assert_includes round_one_html, '<span class="round-result__label--lose">L</span>'
-    assert_includes footer_html, '<div class="footer-score">2 - 0</div>'
-  end
-
-  test "versus header omits team icon when none is attached" do
-    match = create_match_for_renderer_test!(home_name: "Alpha Team", away_name: "Beta Team")
-
-    html = MatchExports::ResultCardRenderer.new(match).send(:versus_html)
-
-    assert_not_includes html, "team-icon"
-  end
-
-  test "versus header embeds team icon as data uri when attached" do
-    match = create_match_for_renderer_test!(home_name: "Alpha Team", away_name: "Beta Team")
-    match.home_team.icon.attach(
-      io: StringIO.new(one_by_one_png),
-      filename: "home_icon.png",
-      content_type: "image/png"
-    )
-
-    html = MatchExports::ResultCardRenderer.new(match.reload).send(:versus_html)
-
-    assert_includes html, %(<img class="team-icon" src="data:image/png;base64,)
-  end
-
   test "last_source_change is nil when no rounds, match_result, or team icons exist" do
     match = create_match_for_renderer_test!(home_name: "Alpha Team", away_name: "Beta Team")
     renderer = MatchExports::ResultCardRenderer.new(match.reload)
